[PORT FROM MAIN] Android Default Camera: disable camera dual open debugging
BZ: 144868

[Description]
Assert() is hit if dual open is detected and the debug flag is being
set to true.

[Implemented Solution]
Disable the DEBUG_OPEN_RELEASE flag and do a error print instead.
Also, the Assert() is relocated so that it will be only hit when
DEBUG_OPEN_RELEASE is enabled, which can be done in development.

Change-Id: I38c30f4b4be0a2899e7e0ff11d2ce02022e5bdb6
Orig-Change-Id: I696bd2508bb88a5b9dc4a046c51f443b820a90a0
Category: feature differentiation
Domain: camera
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Rukai Wei <rukaix.wei@intel.com>